+++ frontend/src/components/orders/hq/HqConfirmedOrderTable.vue
@@ -1,7 +1,7 @@
 <script setup>
 import { ref } from 'vue'
 import { ChevronLeft, ChevronRight } from 'lucide-vue-next'
-import { statusClass, formatPrice } from './orderUtils'
+import { statusClass, formatPrice } from '../orderUtils'
 
 const props = defineProps({
   orders: { type: Array, required: true },
@@ -11,24 +11,15 @@ const props = defineProps({
 
 const emit = defineEmits(['open-detail', 'search', 'page-change'])
 
-const filterType = ref('')
-const dateFrom = ref('')
-const dateTo = ref('')
 const search = ref('')
 
 function emitSearch() {
   emit('search', {
-    ordersType: filterType.value || null,
-    startDate: dateFrom.value || null,
-    endDate: dateTo.value || null,
     keyword: search.value.trim() || null,
   })
 }
 
-function resetFilters() {
-  filterType.value = ''
-  dateFrom.value = ''
-  dateTo.value = ''
+function resetSearch() {
   search.value = ''
   emitSearch()
 }
@@ -37,27 +28,10 @@ function resetFilters() {
 <template>
   <div class="space-y-3">
     <div class="bg-white border border-gray-200 rounded-lg px-5 py-4 flex flex-wrap gap-5 items-end">
-      <label class="flex flex-col gap-2">
-        <span class="text-[10px] font-bold text-gray-400 uppercase tracking-wider">유형</span>
-        <select v-model="filterType"
-          class="w-24 px-3 py-2 rounded border border-gray-200 text-sm outline-none focus:border-[#F37321]">
-          <option value="">전체</option>
-          <option value="AUTO">자동</option>
-          <option value="MANUAL">수동</option>
-        </select>
-      </label>
-      <label class="flex flex-col gap-2">
-        <span class="text-[10px] font-bold text-gray-400 uppercase tracking-wider">기간 시작</span>
-        <input v-model="dateFrom" type="date" class="px-3 py-2 rounded border border-gray-200 text-sm outline-none focus:border-[#F37321]" />
-      </label>
-      <label class="flex flex-col gap-2">
-        <span class="text-[10px] font-bold text-gray-400 uppercase tracking-wider">기간 종료</span>
-        <input v-model="dateTo" type="date" class="px-3 py-2 rounded border border-gray-200 text-sm outline-none focus:border-[#F37321]" />
-      </label>
       <div class="flex flex-col gap-2 flex-1 min-w-40">
-        <span class="text-[10px] font-bold text-gray-400 uppercase tracking-wider">검색</span>
+        <span class="text-[10px] font-bold text-gray-400 uppercase tracking-wider">매장명 검색</span>
         <div class="flex gap-2 items-center">
-          <input v-model="search" type="search" placeholder="발주번호·가맹점"
+          <input v-model="search" type="search" placeholder="매장명을 입력하세요"
             class="flex-1 px-3 py-2 rounded border border-gray-200 text-sm outline-none focus:border-[#F37321]"
             @keyup.enter="emitSearch" />
           <button type="button"
@@ -67,7 +41,7 @@ function resetFilters() {
           </button>
           <button type="button"
             class="text-xs font-semibold text-gray-500 border border-gray-200 px-4 py-2 rounded hover:bg-gray-50 shrink-0 cursor-pointer"
-            @click="resetFilters">
+            @click="resetSearch">
             초기화
           </button>
         </div>
@@ -78,32 +52,26 @@ function resetFilters() {
         <thead>
           <tr class="border-b border-gray-200 bg-gray-50">
             <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">발주번호</th>
-            <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">유형</th>
-            <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">입점 매장</th>
+            <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">가맹점</th>
             <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">발주일시</th>
             <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">총 금액</th>
             <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">상태</th>
           </tr>
         </thead>
         <tbody class="divide-y divide-gray-100">
-          <tr v-for="h in orders" :key="h.id" class="hover:bg-gray-50/50 transition-colors cursor-pointer"
-            @click="$emit('open-detail', h.id)">
-            <td class="px-5 py-3.5 font-mono text-xs text-gray-400">{{ h.id }}</td>
+          <tr v-for="o in orders" :key="o.id"
+            class="hover:bg-gray-50/50 transition-colors cursor-pointer"
+            @click="$emit('open-detail', o.id)">
+            <td class="px-5 py-3.5 font-mono text-xs text-gray-400">{{ o.id }}</td>
+            <td class="px-5 py-3.5 font-semibold text-gray-900">{{ o.store }}</td>
+            <td class="px-5 py-3.5 text-xs text-gray-400 font-mono">{{ o.date }}</td>
+            <td class="px-5 py-3.5 font-semibold text-gray-700">{{ formatPrice(o.price) }}</td>
             <td class="px-5 py-3.5">
-              <span class="text-xs font-bold px-2 py-0.5 rounded"
-                :class="h.type === '자동' ? 'bg-blue-50 text-blue-600 border border-blue-200' : 'bg-purple-50 text-purple-600 border border-purple-200'">
-                {{ h.type }}
-              </span>
-            </td>
-            <td class="px-5 py-3.5 font-semibold text-gray-900">{{ h.store }}</td>
-            <td class="px-5 py-3.5 text-xs text-gray-400 font-mono">{{ h.date }}</td>
-            <td class="px-5 py-3.5 font-semibold text-gray-700">{{ formatPrice(h.price) }}</td>
-            <td class="px-5 py-3.5">
-              <span class="text-xs font-bold px-2 py-0.5 rounded" :class="statusClass(h.status)">{{ h.status }}</span>
+              <span class="text-xs font-bold px-2 py-0.5 rounded" :class="statusClass(o.status)">{{ o.status }}</span>
             </td>
           </tr>
           <tr v-if="orders.length === 0">
-            <td colspan="6" class="px-5 py-10 text-center text-sm text-gray-400">조건에 맞는 발주 이력이 없습니다.</td>
+            <td colspan="5" class="px-5 py-10 text-center text-sm text-gray-400">확정된 발주가 없습니다.</td>
           </tr>
         </tbody>
       </table>
