+++ frontend/src/components/orders/hq/HqDangerSettingsModal.vue
