+++ frontend/src/components/orders/hq/HqOrderHistoryTable.vue
@@ -1,7 +1,7 @@
 <script setup>
 import { ref } from 'vue'
 import { ChevronLeft, ChevronRight } from 'lucide-vue-next'
-import { formatPrice } from './orderUtils'
+import { statusClass, formatPrice } from '../orderUtils'
 
 const props = defineProps({
   orders: { type: Array, required: true },
@@ -9,14 +9,16 @@ const props = defineProps({
   totalPages: { type: Number, default: 1 },
 })
 
-const emit = defineEmits(['open-detail', 'approve', 'reject', 'search', 'page-change'])
+const emit = defineEmits(['open-detail', 'search', 'page-change'])
 
+const filterType = ref('')
 const dateFrom = ref('')
 const dateTo = ref('')
 const search = ref('')
 
 function emitSearch() {
   emit('search', {
+    ordersType: filterType.value || null,
     startDate: dateFrom.value || null,
     endDate: dateTo.value || null,
     keyword: search.value.trim() || null,
@@ -24,6 +26,7 @@ function emitSearch() {
 }
 
 function resetFilters() {
+  filterType.value = ''
   dateFrom.value = ''
   dateTo.value = ''
   search.value = ''
@@ -34,6 +37,15 @@ function resetFilters() {
 <template>
   <div class="space-y-3">
     <div class="bg-white border border-gray-200 rounded-lg px-5 py-4 flex flex-wrap gap-5 items-end">
+      <label class="flex flex-col gap-2">
+        <span class="text-[10px] font-bold text-gray-400 uppercase tracking-wider">유형</span>
+        <select v-model="filterType"
+          class="w-24 px-3 py-2 rounded border border-gray-200 text-sm outline-none focus:border-[#F37321]">
+          <option value="">전체</option>
+          <option value="AUTO">자동</option>
+          <option value="MANUAL">수동</option>
+        </select>
+      </label>
       <label class="flex flex-col gap-2">
         <span class="text-[10px] font-bold text-gray-400 uppercase tracking-wider">기간 시작</span>
         <input v-model="dateFrom" type="date" class="px-3 py-2 rounded border border-gray-200 text-sm outline-none focus:border-[#F37321]" />
@@ -66,52 +78,32 @@ function resetFilters() {
         <thead>
           <tr class="border-b border-gray-200 bg-gray-50">
             <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">발주번호</th>
+            <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">유형</th>
             <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">입점 매장</th>
-            <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">발주수량</th>
-            <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">평균수량</th>
-            <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">초과율</th>
-            <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">총 금액</th>
             <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">발주일시</th>
+            <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">총 금액</th>
             <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">상태</th>
-            <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider text-center">처리</th>
           </tr>
         </thead>
         <tbody class="divide-y divide-gray-100">
-          <tr v-for="o in orders" :key="o.id"
-            class="hover:bg-gray-50/50 transition-colors cursor-pointer"
-            @click="$emit('open-detail', o.id)"
-          >
-            <td class="px-5 py-3.5 font-mono text-xs text-gray-400">{{ o.id }}</td>
-            <td class="px-5 py-3.5 font-semibold text-gray-900">{{ o.store }}</td>
-            <td class="px-5 py-3.5 font-bold text-red-600">{{ (o.qty ?? 0).toLocaleString() }}</td>
-            <td class="px-5 py-3.5 text-gray-500">{{ (o.avgQty ?? 0).toLocaleString() }}</td>
-            <td class="px-5 py-3.5">
-              <span class="text-xs font-black px-2 py-0.5 rounded bg-red-50 text-red-600 border border-red-200">
-                +{{ o.ratio }}%
-              </span>
-            </td>
-            <td class="px-5 py-3.5 font-semibold text-gray-700">{{ formatPrice(o.price) }}</td>
-            <td class="px-5 py-3.5 text-xs text-gray-400 font-mono">{{ o.date }}</td>
+          <tr v-for="h in orders" :key="h.id" class="hover:bg-gray-50/50 transition-colors cursor-pointer"
+            @click="$emit('open-detail', h.id)">
+            <td class="px-5 py-3.5 font-mono text-xs text-gray-400">{{ h.id }}</td>
             <td class="px-5 py-3.5">
               <span class="text-xs font-bold px-2 py-0.5 rounded"
-                :class="o.status === 'APPROVE' || o.status === 'REJECT'
-                  ? 'bg-gray-100 text-gray-400 border border-gray-200'
-                  : 'bg-red-50 text-red-600 border border-red-200'">
-                {{ o.status === 'APPROVE' || o.status === 'REJECT' ? '처리완료' : 'DANGER' }}
+                :class="h.type === '자동' ? 'bg-blue-50 text-blue-600 border border-blue-200' : 'bg-purple-50 text-purple-600 border border-purple-200'">
+                {{ h.type }}
               </span>
             </td>
+            <td class="px-5 py-3.5 font-semibold text-gray-900">{{ h.store }}</td>
+            <td class="px-5 py-3.5 text-xs text-gray-400 font-mono">{{ h.date }}</td>
+            <td class="px-5 py-3.5 font-semibold text-gray-700">{{ formatPrice(h.price) }}</td>
             <td class="px-5 py-3.5">
-              <div v-if="o.status !== 'APPROVE' && o.status !== 'REJECT'" class="flex justify-center gap-1.5">
-                <button @click.stop="$emit('approve', o)"
-                  class="px-2.5 py-1 bg-[#F37321] text-white text-xs font-semibold hover:bg-[#e0661d] rounded cursor-pointer">승인</button>
-                <button @click.stop="$emit('reject', o)"
-                  class="px-2.5 py-1 border border-gray-200 text-gray-600 text-xs font-medium hover:bg-gray-50 rounded cursor-pointer">반려</button>
-              </div>
-              <span v-else class="text-xs text-gray-400 block text-center">—</span>
+              <span class="text-xs font-bold px-2 py-0.5 rounded" :class="statusClass(h.status)">{{ h.status }}</span>
             </td>
           </tr>
           <tr v-if="orders.length === 0">
-            <td colspan="9" class="px-5 py-10 text-center text-sm text-gray-400">이상 발주가 없습니다.</td>
+            <td colspan="6" class="px-5 py-10 text-center text-sm text-gray-400">조건에 맞는 발주 이력이 없습니다.</td>
           </tr>
         </tbody>
       </table>
