+++ frontend/src/components/orders/store/StoreItemDeleteModal.vue
